requirements.txt
asgiref==3.9.1 ; python_version >= "3.13" and python_version < "4.0"
certifi==2025.7.14 ; python_version >= "3.13" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.13" and python_version < "4.0" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.2 ; python_version >= "3.13" and python_version < "4.0"
cryptography==45.0.5 ; python_version >= "3.13" and python_version < "4.0"
defusedxml==0.7.1 ; python_version >= "3.13" and python_version < "4.0"
django-environ==0.12.0 ; python_version >= "3.13" and python_version < "4.0"
django==5.2.4 ; python_version >= "3.13" and python_version < "4.0"
idna==3.10 ; python_version >= "3.13" and python_version < "4.0"
oauthlib==3.3.1 ; python_version >= "3.13" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.13" and python_version < "4.0" and platform_python_implementation != "PyPy"
pyjwt==2.10.1 ; python_version >= "3.13" and python_version < "4.0"
python3-openid==3.2.0 ; python_version >= "3.13" and python_version < "4.0"
requests-oauthlib==2.0.0 ; python_version >= "3.13" and python_version < "4.0"
requests==2.32.4 ; python_version >= "3.13" and python_version < "4.0"
social-auth-app-django==5.5.1 ; python_version >= "3.13" and python_version < "4.0"
social-auth-core==4.7.0 ; python_version >= "3.13" and python_version < "4.0"
sqlparse==0.5.3 ; python_version >= "3.13" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.13" and python_version < "4.0" and sys_platform == "win32"
urllib3==2.5.0 ; python_version >= "3.13" and python_version < "4.0"
